+++ backend/motoextreme/motoextreme/src/main/java/com/motoextreme/motoextreme/services/UsuarioService.java
@@ -2,6 +2,7 @@ package com.motoextreme.motoextreme.services;
 
 import com.motoextreme.motoextreme.dtos.request.UsuarioRequestDTO;
 import com.motoextreme.motoextreme.dtos.response.UsuarioResponseDTO;
+import com.motoextreme.motoextreme.exeptions.ResourceNotFoundExeption;
 import com.motoextreme.motoextreme.mappers.UsuarioMapper;
 import com.motoextreme.motoextreme.models.entities.Carrito;
 import com.motoextreme.motoextreme.models.entities.Usuario;
@@ -12,7 +13,6 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-
 @Service
 @RequiredArgsConstructor
 public class UsuarioService {
@@ -20,7 +20,7 @@ public class UsuarioService {
     private final IUsuario usuarioRepository;
     private final UsuarioMapper usuarioMapper;
 
-   //listar usuarios
+    // Listar usuarios
     public List<UsuarioResponseDTO> findAll() {
         List<Usuario> usuarios = usuarioRepository.findAll();
         return usuarios.stream()
@@ -28,36 +28,36 @@ public class UsuarioService {
                 .toList();
     }
 
-    //crear usuario y asignarle un carrito
+    // Crear usuario y asignarle carrito
     public UsuarioResponseDTO crearUsuario(UsuarioRequestDTO dto) {
-
         Usuario usuario = usuarioMapper.toEntity(dto);
 
         Carrito carrito = new Carrito();
         carrito.setUsuario(usuario);
         usuario.setCarrito(carrito);
 
-
         Usuario guardado = usuarioRepository.save(usuario);
-
         return usuarioMapper.toDto(guardado);
     }
 
-    //buscar por id
+    // Buscar por id
     public Optional<UsuarioResponseDTO> findById(Long id) {
         Optional<Usuario> usuario = usuarioRepository.findById(id);
         return usuario.map(usuarioMapper::toDto);
     }
 
-   //eliminar usuario
+    // Eliminar usuario
     public void eliminarUsuario(Long id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new ResourceNotFoundExeption("Usuario no encontrado");
+        }
         usuarioRepository.deleteById(id);
     }
 
-   //actualizar usuario
+    // Actualizar usuario
     public UsuarioResponseDTO actualizarUsuario(Long id, UsuarioRequestDTO dto) {
         Usuario usuario = usuarioRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundExeption("Usuario no encontrado"));
 
         usuario.setNombre(dto.getNombre());
         usuario.setEmail(dto.getEmail());
@@ -65,9 +65,6 @@ public class UsuarioService {
         usuario.setRol(dto.getRol());
 
         Usuario actualizado = usuarioRepository.save(usuario);
-
         return usuarioMapper.toDto(actualizado);
     }
-
-
 }
